% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_EI30.R
\name{event.EI30}
\alias{event.EI30}
\title{Rainfall Erosivity Index}
\usage{
event.EI30(
  data,
  precipitation,
  temperature,
  event = "Erosive_event",
  data.I30,
  I30.event = "Erosive_event",
  I30.label = "I30_mm.h.1",
  record.step = "00:10:00",
  ceiling = FALSE,
  precip.units = "mm",
  min.temperature = 0,
  digits = 0,
  ir.factor = 0.08
)
}
\arguments{
\item{data}{A data frame containing the precipitation data.}

\item{precipitation}{A string specifying the name of the column in `data` that contains the precipitation values.}

\item{temperature}{A string specifying the name of the column in `data` that contains the temperature values (in Celsius).}

\item{event}{The name of the column in `data` containing event labels.}

\item{data.I30}{Optional: A data frame containing pre-calculated I30 values for each event. If not provided, the function will calculate I30 values.}

\item{I30.event}{The name of the column in `data.I30` containing event labels.}

\item{I30.label}{The name of the column in `data.I30` containing I30 values.}

\item{record.step}{A string representing the time step of the precipitation records in the format "HH:MM:SS". Default is "00:10:00".}

\item{ceiling}{A logical value indicating whether to use the ceiling function when calculating periods. Default is `FALSE.`}

\item{precip.units}{A string specifying the units of precipitation. Acceptable values are "mm" (default), "cm", and "m".}

\item{min.temperature}{A numeric value specifying the minimum temperature required for a precipitation to be considered as snow and not precipitation. Default is 0.}

\item{digits}{An integer specifying the number of decimal places to round the EI30 values. Default is 0.}

\item{ir.factor}{An integer specifying the value to which the precipitation intensity for each time interval should be multiplicated to. In Renard and Freimund (1994) = 0.04 and in RUSLE2 = 0.08 (default; Yin et al., 2017)}
}
\value{
A data frame with the calculated EI30 values for each erosive event.
}
\description{
This function calculates the EI30 (Rainfall Erosivity Index) for each erosive event in a dataset. It uses precipitation data and optionally pre-calculated I30 values to compute the index.
}
\details{
The function calculates how many times the record step should be multiplied to reach one hour. It then corrects the precipitation units if necessary. For each event, it calculates the rainfall energy and uses pre-calculated or newly calculated I30 values to compute the EI30 index.
}
\references{
Renard, K. G., & Freimund, J. R. (1994). Using monthly precipitation data to estimate the r-factor in the revised USLE, . 157 , 287–306. URL: https://linkinghub.elsevier.com/retrieve/pii/0022169494901104. doi: 10.1016/0022-1694(94)90110-

Yin, S., Nearing, M. A., Borrelli, P., Xue, X. (2017). Rainfall Erosivity: An Overview of Methodologies and Applications, Vadose Zone Journal, 16 (12) , 1-16. DOI: https://acsess.onlinelibrary.wiley.com/doi/10.2136/vzj2017.06.0131. doi:10.2136/vzj2017.06.0131
}
\author{
Thomas Chalaux-Clergue
}
