% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_I30.R
\name{event.I30}
\alias{event.I30}
\title{Calculate the Maximum 30-Minute Intensity for Erosive Events}
\usage{
event.I30(
  data,
  precipitation,
  temperature,
  event = "Erosive_event",
  record.step = "00:10:00",
  ceiling = FALSE,
  min.temperature = 0,
  precip.units = "mm"
)
}
\arguments{
\item{data}{A data frame containing the precipitation data.}

\item{precipitation}{A string specifying the name of the column in `data` that contains the precipitation values.}

\item{temperature}{A string specifying the name of the column in `data` that contains the temperature values (in Celsius).}

\item{record.step}{A string specifying the time step between records in the format "HH:MM:SS". Default is "00:10:30".}

\item{ceiling}{A logical value indicating whether to use the ceiling function when calculating periods. Default is `FALSE`.}

\item{min.temperature}{A numeric value specifying the minimum temperature required for a precipitation to be considered as snow and not precipitation. Default is 0.}

\item{precip.units}{A string specifying the units of precipitation. Acceptable values are "mm" (default), "cm", and "m".}

\item{events}{A string specifying the name of the column in `data` that contains the erosive event labels. Default is "Erosive_Event".}
}
\value{
A data frame with columns for the event labels and the corresponding maximum 30-minute precipitation intensity (I30) in mm/h.
}
\description{
This function calculates the maximum 30-minute precipitation intensity (I30) for each erosive event in a dataset.
}
\details{
The function calculates the maximum 30-minute precipitation intensity (I30) for each erosive event in the dataset. The time step between records is specified by `record.step`, and the units of precipitation can be specified using `precip.units`.
If `ceiling` is TRUE, the ceiling function is used to calculate the number of periods needed for 30 minutes; otherwise, the round function is used.
}
\examples{
\dontrun{
data <- data.frame(date = seq.POSIXt(from = as.POSIXct("2023-01-01 00:00:00"),
                                     to = as.POSIXct("2023-01-01 03:30:00"),
                                     by = "10 min"),
                   rainfall = c(0, 0.5, 0.2, 0.1, 0, 0, 0, 0, 0, 0, 0, 2, 3, 0, 0, 1, 0, 0, 0, 0, 0, 0),
                   Event = c(NA, 1, 1, 1, NA, NA, NA, NA, NA, NA, NA, 2, 2, 2, 2, 2, NA, NA, NA, NA, NA, NA)
                   Erosive_Event = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1, 1, 1, 1, 1, NA, NA, NA, NA, NA, NA))
event.I30(data, precipitation = "rainfall")
}

}
\references{
Renard, K. G., & Freimund, J. R. (). Using monthly precipitation data to estimate the r-factor in the revised USLE, . 157 , 287–306. URL: https://linkinghub.elsevier.com/retrieve/pii/0022169494901104. doi:10.1016/0022-1694(94)90110-
}
\author{
Thomas Chalaux-Clergue
}
