% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_splitter.R
\name{event.splitter}
\alias{event.splitter}
\title{Split Precipitation Data into Events}
\usage{
event.splitter(
  data,
  precipitation,
  record.step = "00:10:00",
  delay.between.events = "06:00:00",
  min.bw.events = 1.27,
  ceiling = FALSE
)
}
\arguments{
\item{data}{A data frame containing the precipitation data.}

\item{precipitation}{A string specifying the name of the column in `data` that contains the precipitation values.}

\item{record.step}{A string specifying the time step between consecutive records in the format "HH:MM:SS". Default is "00:10:00".}

\item{delay.between.events}{A string specifying the maximum allowable delay between consecutive precipitation events in the format "HH:MM:SS". Default is "06:00:00".}

\item{min.bw.events}{A numeric value specifying the minimum precipitation required to separate two events. Default is 1.27 mm.}

\item{ceiling}{A logical value indicating whether to use the ceiling function when calculating periods. Default is `FALSE`.}
}
\value{
A data frame containing the original data with an additional column named "Event" indicating the event number for each record.
}
\description{
This function splits a precipitation dataset into separate events based on specified criteria.
An event is defined as a series of precipitation records with no more than a specified delay
between them and with a minimum precipitation between events.
}
\details{
The function labels each precipitation record with an event number, indicating which event it belongs to.
An event is considered to have ended if there is a period of time longer than `delay.between.events` with
no precipitation or if the sum of precipitation during this period is less than `min.bw.events`.
}
\examples{
\dontrun{
data <- data.frame(date = seq.POSIXt(from = as.POSIXct("2023-01-01 00:00:00"),
                                     to = as.POSIXct("2023-01-02 00:00:00"),
                                     by = "10 min"),
                   rainfall = c(0, 0.5, 0.2, 0, 0, 0, 0, 0, 0, 0, 2, 3, 0, 0, 1, 0, 0, 0, 1, 0.5, 0))
event.splitter(data, precipitation = "rainfall")
}

}
\references{
Meusburger, K., Steel, A., Panagos, P., Montanarella, L., & Alewell, C. (2012). Spatial and temporal variability of rainfall erosivity factor for switzerland, . 16 , 167–177. URL: https://hess.copernicus.org/articles/16/167/2012/. doi:10.5194/hess-16-167-2012.
}
\author{
Thomas Chalaux-Clergue
}
