% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_summary.R
\name{event.summary}
\alias{event.summary}
\title{Summarize Erosive Precipitation Events}
\usage{
event.summary(
  data,
  precipitation,
  record.step = "00:10:00",
  delay.between.event = "06:00:00",
  temperature,
  Date,
  min.bw.events = 1.27,
  min.precipitation = 12.7,
  ceiling = FALSE,
  adapt.label = TRUE,
  precip.units = "mm",
  min.temperature = 0,
  digits = 0,
  ir.factor = 0.08,
  save.dir,
  note
)
}
\arguments{
\item{data}{A data frame containing precipitation data.}

\item{precipitation}{A string specifying the name of the column in `data` that contains the precipitation values.}

\item{record.step}{A string specifying the time step between consecutive records in the format "HH:MM:SS". Default is "00:10:00".}

\item{delay.between.event}{A string specifying the maximum allowable delay between consecutive precipitation events in the format "HH:MM:SS". Default is "06:00:00".}

\item{temperature}{A string specifying the name of the column in `data` that contains the temperature values (in Celsius).}

\item{Date}{A string specifying the column name in `data` representing the date and time of each record.}

\item{min.bw.events}{A numeric value specifying the minimum precipitation required to separate two events. Default is 1.27 mm.}

\item{min.precipitation}{A numeric value specifying the minimum cumulative precipitation required for an event to be considered erosive. Default is 12.7.}

\item{ceiling}{A logical value indicating whether to use the ceiling function when calculating periods. Default is `FALSE`.}

\item{adapt.label}{A logical value indicating whether to update the event label after filtering by minimum precipitation. Default is `TRUE`.}

\item{precip.units}{A string specifying the units of precipitation. Acceptable values are "mm" (default), "cm", and "m".}

\item{min.temperature}{A numeric value indicating the minimum temperature to consider for precipitation (e.g., to exclude snow). Default is 0.}

\item{digits}{An integer specifying the number of decimal places to round the EI30 values. Default is 0.}

\item{ir.factor}{An integer specifying the value to which the precipitation intensity for each time interval should be multiplicated to. In Renard and Freimund (1994) = 0.04 and in RUSLE2 = 0.08 (default; Yin et al., 2017)}

\item{save.dir}{A optional Connection open for writing the test results data.frame. If "" save the file at working directory.}

\item{note}{A optional character string to add a note at the end of the file name.}
}
\value{
A data frame summarizing each identified erosive precipitation event, including:
\describe{
  \item{Event_label}{The label of the event.}
  \item{Date_start}{The start date and time of the event.}
  \item{Date_end}{The end date and time of the event.}
  \item{Duration_min}{The duration of the event in minutes (min).}
  \item{Duration_h}{The duration of the event in hours (h).}
  \item{Cumulative_precipitation_mm}{The cumulative precipitation of the event in millimeters (mm).}
  \item{I30_mm.h.1}{The maximum 30-minute intensity (I30) of the event in mm h^{-1}.}
  \item{EI30_MJ.mm.ha.1.h.1}{The Rainfall Erosivity Index (EI30) of the event in MJ mm ha^{-1} h^{-1}.}
}
}
\description{
This function identifies and summarizes erosive precipitation events in a dataset, calculating key metrics such as cumulative precipitation, event duration, maximum 30-minute intensity (I30), and Rainfall Erosivity Index (EI30).
}
\author{
Thomas Chalaux-Clergue
}
