% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_thresholder.R
\name{event.thresholder}
\alias{event.thresholder}
\title{Filter Erosive Precipitation Events}
\usage{
event.thresholder(
  data,
  precipitation,
  event = "Event",
  min.precipitation = 12.7,
  adapt.label = TRUE
)
}
\arguments{
\item{data}{A data frame containing the precipitation data.}

\item{precipitation}{A string specifying the name of the column in `data` that contains the precipitation values.}

\item{min.precipitation}{A numeric value specifying the minimum cumulative precipitation required for an event to be considered erosive. Default is 12.7.}

\item{adapt.label}{A logical value indicating whether to update the event label after filtering by minimum precipitation. Default is `TRUE`.}

\item{events}{A string specifying the name of the column in `data` that contains the event labels. Default is "Events".}
}
\value{
A data frame containing the original data with an additional column named "Erosive_events" indicating the erosive event number for each record.
}
\description{
This function identifies and labels precipitation events that exceed a specified minimum cumulative precipitation threshold.
}
\details{
The function filters precipitation events based on the cumulative precipitation threshold and labels the events that exceed this threshold.
If `adapt.label` is TRUE, the identified events will be re-labelled in sequential order. Otherwise, the original event labels are retained.
}
\examples{
\dontrun{
data <- data.frame(date = seq.POSIXt(from = as.POSIXct("2023-01-01 00:00:00"),
                                     to = as.POSIXct("2023-01-01 03:30:00"),
                                     by = "10 min"),
                   rainfall = c(0, 1.5, 2.2, 2, 0, 0, 0, 0, 0, 0, 0, 2, 3, 0, 0, 1, 0, 0, 0, 0, 0, 0),
                   Event = c(NA, 1, 1, 1, NA, NA, NA, NA, NA, NA, NA, 2, 2, 2, 2, 2, NA, NA, NA, NA, NA, NA))
event.thresholder(data, precipitation = "rainfall")
}

}
\references{
Meusburger, K., Steel, A., Panagos, P., Montanarella, L., & Alewell, C. (2012). Spatial and temporal variability of rainfall erosivity factor for switzerland, . 16 , 167–177. URL: https://hess.copernicus.org/articles/16/167/2012/. doi:10.5194/hess-16-167-2012.
}
\author{
Thomas Chalaux-Clergue
}
